New command logic tests.
